chore(egui): bump to 0.33.3 (egui/eframe/extras/phosphor/catppuccin)
- workspace deps: egui/eframe/egui_extras 0.29 → 0.33.3；catppuccin-egui
  5 egui29 → 5.7 egui33；egui-phosphor 0.7 → 0.11
- eframe features 不再 default-features = false（wgpu backend 要靠 default
  feature 选择，否则 wgpu 27 panic "No wgpu backend feature ..."）
- API 迁移:
  - Frame::none() → Frame::new()
  - Frame::rounding(f) → Frame::corner_radius(f)
  - Margin::symmetric(f32, f32) → (i8, i8)，全部加 `as i8`
  - Margin { left/right/top/bottom: f32 } → i8
  - Visuals::window_rounding → window_corner_radius
  - Visuals::menu_rounding → menu_corner_radius
  - WidgetVisuals::rounding 字段 → corner_radius
  - FontData::from_owned(bytes) 包 std::sync::Arc::new(...)
- 测试全绿；本地启动无 panic

+++ crates/modbusmaster-egui/src/app.rs
@@ -980,7 +980,7 @@ impl eframe::App for MasterApp {
                 ui,
                 flavor,
                 theme::Layer::L1,
-                egui::Margin::symmetric(14.0, 10.0),
+                egui::Margin::symmetric(14.0 as i8, 10.0 as i8),
                 |ui| {
                     ui.horizontal(|ui| {
                         ui.label(egui::RichText::new(&s.label).strong().size(13.5));
@@ -996,7 +996,7 @@ impl eframe::App for MasterApp {
             );
             ui.add_space(4.0);
 
-            uikit::region(ui, flavor, theme::Layer::L1, egui::Margin::symmetric(14.0, 10.0), |ui| {
+            uikit::region(ui, flavor, theme::Layer::L1, egui::Margin::symmetric(14.0 as i8, 10.0 as i8), |ui| {
             // Tab bar: Read / Write / Poll
             ui.horizontal(|ui| {
                 for tab in [MasterTab::Read, MasterTab::Write, MasterTab::Poll] {
@@ -1247,7 +1247,7 @@ impl eframe::App for MasterApp {
                 .unwrap_or((None, 0));
             let show_result = poll_latest.clone().or_else(|| self.read_result.clone());
             if let Some(result) = &show_result {
-                uikit::region(ui, flavor, theme::Layer::L2, egui::Margin::symmetric(12.0, 10.0), |ui| {
+                uikit::region(ui, flavor, theme::Layer::L2, egui::Margin::symmetric(12.0 as i8, 10.0 as i8), |ui| {
                 let title = if poll_latest.is_some() { "轮询结果" } else { "读取结果" };
                 let base = if poll_latest.is_some() { poll_addr } else { self.read_addr };
                 ui.label(egui::RichText::new(title).strong().size(12.5));
